Prevent calling unbind on undefined firebaseRefs
Currently the componentWillUnmount will attempt to unbind a ref that we've already called unbind on, resulting in an error.
We need to skip any values within this.firebaseRefs that equal "undefined". Alternatively we could change unbind to delete the ref from this.firebaseRefs (it looks like it used to be that way and was changed for some reason).